app/backend/requirements.txt
azure-identity==1.13.0
quart==0.19.0
langchain>=0.0.187
openai[datalib]==0.27.8
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
uvicorn[standard]==0.23.2
aiohttp==3.8.5
azure-monitor-opentelemetry==1.0.0b15
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
python-jose==3.3.0
azure-cosmos==4.5.0
quart-schema==0.17.1
pypdf==3.9.0
azure-ai-formrecognizer==3.2.1
tenacity==8.2.2
multimethod==1.10